remove begin and end from default filters; add self.filters to DeviceCollection
there are sensible defaults for begin and end on the server side.  specifying
them here can cause problems when trying to retrieve lists of interfaces: the
server default is to return only currently active interfaces but if you
specifyu a time range it will get all interfaces in that time range.

This also brings up an issue that while it is possible to change defaults
there is no way currently to unset a default in ApiFilters. Thus, if you want to
revert to the server default you can't.