do not insert null/empty rows when there is a gap
We now log gaps that exceed the heartbeat threshold but do not backfill with
null/empty rows.

Backfilling gaps has caused data loss in the rare situation that we get data
into the persister out of order. The intentions of this backfilling are lost
in the mists of time and the approach seems dubious as it might casuse data to
be lost.