Two changes in this, one complete and one partial.
GenericOidCorrelator takes as an argument the name of an OID to
use as the mapping from index to textual description. So for ifTable
you might use ifName or ifDescr. The Poller args are now passed to
the __init__ of all the Correlators with a default of None.

*** The Correlator should check if multiple indexes are mapping to
one key and either raise an exception or work around it by for
example adding the raw index onto each key.
***

The RowRef and GenericTableRef are works in progress. They will be used
eventually to store whatever correlator_oid the GenericOidCorrelator is
configured for so that the API can make the collected data browsable.